feat: remove connected react router
