--- src/modules/location/index.ts
@@ -1,2 +1,0 @@
-export * from './sagas'
-export * from './types'
