--- src/modules/location/types.ts
@@ -1,10 +1,0 @@
-import { Network } from '@dcl/schemas/dist/dapps/network'
-import { NetworkGatewayType } from 'decentraland-ui'
-import { MoonPayTransactionStatus } from '../gateway/moonpay/types'
-
-export type RedirectTransactionQuery = {
-  network?: Network
-  gateway?: NetworkGatewayType
-  transactionId?: string
-  transactionStatus?: MoonPayTransactionStatus
-}
